requirements.txt
requests
trafilatura
google-generativeai
beautifulsoup4